Move parse.py to client submodule
